New completion method
Robot can now detect corners and turn appropriatly. It will not turn right unless there is no path infront of the robot so the robot always follows the straight path or takes a left, should be able to deal with T intersections and dead ends better, and should be able to deal with plus intersections now. The code does need to be compressed though as it is well past the 80 line per function limit