First Challenge Iteration
Working challenge code. Brought ideas from the previous version and combinded them. This challenge code works by hugging the left wall as tight as possible, if it encounters a wall infront of itm it does a 90 degree turn to the right, and if it can't find the left wall it keeps turning left until it finds it. For alignment with the wall, it also checks a shifted position of the wall and if the wall is in the shifted position turn right to correct  the position.