Core, Completion, Challenge fix
Tested the program on core and completion and I came across some errors (not stopping on the flag, not doing 90 degree turns right, and instantly going into maze mode when the robot can't find any white). The fix for the flag was to make a new function to detect black pixels in an area. The 90 degree right was an issue with the cleaned up code by Samuel where he forgot to set turningLeft and turningRight when the robot didn't need to turn left or right. And finally the instant maze mode when no white could be found was my fault as I was doing if(redCheck) where it should be if(redCheck()) as the first option always was true